graphql/requirements.txt
fastapi
uvicorn
strawberry-graphql
sqlalchemy
asyncpg
psycopg2-binary
